Working 1 flag ctf
Added capture zone to the middle of the map. Players can only have 1
flag, running over a flag with one won't pick it up. Running into the
green light in the middle with a flag counts as capping one. Players
with a flag have a purple light above them. Lights are slightly hard to
see - directional light is too bright maybe (the spotlights are on max
intensity and still barely visible)?